+++ src/_analytics_google_script.d.ts
@@ -1,6 +1,6 @@
-import type { gtm_body_script_html_opts_type } from './gtm_body_script_html_opts_type';
+import type { gtm_body_script_html_opts_T } from './gtm_body_script_html_opts_T';
 /**
  * Html to add ga.js to the site
  */
-export declare function _analytics_google_script(...opts_a1: Partial<gtm_body_script_html_opts_type>[]): string;
+export declare function _analytics_google_script(...opts_a1: Partial<gtm_body_script_html_opts_T>[]): string;
 export { _analytics_google_script as _script__google__analytics };