+++ src/_analytics_google_script.ts
@@ -1,11 +1,11 @@
 import { clone } from '@ctx-core/object'
 import { error_Ctx, throw_missing_argument } from '@ctx-core/error'
-import type { gtm_body_script_html_opts_type } from './gtm_body_script_html_opts_type'
+import type { gtm_body_script_html_opts_T } from './gtm_body_script_html_opts_T'
 /**
  * Html to add ga.js to the site
  */
-export function _analytics_google_script(...opts_a1:Partial<gtm_body_script_html_opts_type>[]) {
-	const opts = clone(...opts_a1) as gtm_body_script_html_opts_type
+export function _analytics_google_script(...opts_a1:Partial<gtm_body_script_html_opts_T>[]) {
+	const opts = clone(...opts_a1) as gtm_body_script_html_opts_T
 	const GOOGLE_TRACKING_ID = opts.GOOGLE_TRACKING_ID || opts.GA_ID || process.env.GA_GOOGLE_TRACKING_ID
 	if (!GOOGLE_TRACKING_ID) throw_missing_argument({ key: 'process.env.GOOGLE_TRACKING_ID' } as error_Ctx)
 	return `