+++ src/_gtm_body_script_html.d.ts
@@ -1,6 +1,6 @@
-import type { gtm_body_script_html_opts_type } from './gtm_body_script_html_opts_type';
+import type { gtm_body_script_html_opts_T } from './gtm_body_script_html_opts_T';
 /**
  * Google Tag Manager script html to place at the top of `<body>`
  */
-export declare function _gtm_body_script_html(opts?: gtm_body_script_html_opts_type): string;
+export declare function _gtm_body_script_html(opts?: gtm_body_script_html_opts_T): string;
 export { _gtm_body_script_html as _html__script__body__gtm };