+++ src/_gtm_script_html.ts
@@ -1,4 +1,4 @@
-import type { _gtm_script_html_opts_type } from './_gtm_script_html_opts_type'
+import type { _gtm_script_html_opts_T } from './_gtm_script_html_opts_T'
 import { _gtm_head_script_html } from './_gtm_head_script_html'
 import { _gtm_body_script_html } from './_gtm_body_script_html'
 /**
@@ -7,7 +7,7 @@ import { _gtm_body_script_html } from './_gtm_body_script_html'
  * @param opts.GTM_ID
  * @returns {string}
  */
-export function _gtm_script_html(opts = {} as _gtm_script_html_opts_type) {
+export function _gtm_script_html(opts = {} as _gtm_script_html_opts_T) {
 	return `
 ${_gtm_head_script_html(opts)}
 ${_gtm_body_script_html(opts)}