+++ src/_gtm_script_html_opts_T.d.ts
@@ -1,5 +1,5 @@
 import type { error_Ctx } from '@ctx-core/error';
-export interface _gtm_script_html_opts_type extends error_Ctx {
+export interface _gtm_script_html_opts_T extends error_Ctx {
     GTM_ID?: string;
     dataLayer?: [];
 }